Add metric 'rabbitmq_up'
The new metric exposes whether the last scrape of RabbitMQ was
successful. It is inspired by the haproxy_exporter which exposes the
metric `haproxy_up`.
The value of `rabbitmq_up` is either `1` or `0`. It can be used to
determine whether the RabbitMQ process being scraped is up-and-running.

In order to be able to set the metric, some functions have been
rewritten to return error values instead of using callbacks.